Chore: supprime les fichiers pyc générés
