+++ Start.py
@@ -1,0 +1,13 @@
+import tkinter as tk
+
+from modules.main_app import MainApp
+
+
+def main() -> None:
+    root = tk.Tk()
+    MainApp(root)
+    root.mainloop()
+
+
+if __name__ == "__main__":
+    main()
